version updated backend

+++ app/Controllers/Http/CourseController.js
@@ -7,7 +7,8 @@
 const Course = use( 'App/Models/Course' )
 const Enterprise = use( 'App/Models/Enterprise' )
 const User = use( 'App/Models/User' )
-const Databsae = use( 'Database' )
+const Database = use( 'Database' )
+
 class CourseController {
   /**
    * Show a list of all courses.
@@ -18,8 +19,11 @@ class CourseController {
    * @param {Response} ctx.response
    * @param {View} ctx.view
    */
-  async index ( { request, response, view } ) {
-    const courses = await Course.query().with( 'enterprise' ).fetch()
+  async index ( { params } ) {
+    const courses = await Course.query()
+      .where( 'enterprise_id', params.enterprise_id )
+      .with( 'enterprise' )
+      .fetch()
 
     return courses
   }
@@ -32,16 +36,16 @@ class CourseController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async store ( { request, auth } ) {
-    const data = request.only( ['title', 'description'] )
-
-    const id = await Databsae
-      .select( 'id' )
-      .table( 'enterprises' )
-      .where( { user_id: auth.user.id } )
-
+  async store ( { request, auth, params, response } ) {
+    const data = request.only( [
+      'title',
+      'description',
+      'nicho',
+      'due_date',
+      'person_id'
+    ] )
     const course = await Course
-      .create( { ...data, enterprise_id: id[0].id } )
+      .create( { ...data, enterprise_id: params.enterprise_id } )
 
     return course
   }
@@ -55,12 +59,17 @@ class CourseController {
    * @param {Response} ctx.response
    * @param {View} ctx.view
    */
-  async show ( { params, request, response, view } ) {
-    const course = await Course.findOrFail( params.id )
+  async show ( { params } ) {
+    try {
+      const course = await Course.findOrFail( params.id )
 
-    await course.load( 'enterprise' )
+      await course.load( 'enterprise' )
 
-    return course
+      return course
+
+    } catch ( error ) {
+      console.log( `Same wrong on: ${error}` );
+    }
   }
 
 
@@ -73,14 +82,37 @@ class CourseController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async update ( { params, request, response } ) {
+  async update ( { params, request, auth } ) {
     const course = await Course.findOrFail( params.id )
-    const data = request.only( ['title', 'description'] )
+    let data = request.only( [
+      'title',
+      'description',
+      'nicho',
+      'due_date',
 
+    ] )
+    const user_auth = await auth.getUser()
+    const user = await User.findOrFail( user_auth.id )
+
+    if ( user.isEnterprise === '0' ) {
+      const person = await Database.select( 'id', 'etnia' )
+        .from( 'people' )
+        .where( 'user_id', user_auth.id )
+
+      /*       if ( person.etnia !== course.nicho ) {
+              return
+            } */
+
+      let value
+      if ( course.person_id === person[0].id ) {
+        value = { person_id: null }
+      } else {
+        value = { person_id: person[0].id }
+      }
+      data = value
+    }
     course.merge( data )
-
     await course.save()
-
     return course
   }
 
@@ -92,10 +124,12 @@ class CourseController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async destroy ( { params, request, response } ) {
+  async destroy ( { params, response } ) {
     const course = await Course.findOrFail( params.id )
-
-    await course.delete()
+    const enterprise = await Enterprise.findOrFail( course.enterprise_id )
+    if ( course.enterprise_id === enterprise.id )
+      return await course.delete()
+    return response.send( 'Something wrong' )
   }
 }
